Remove public domain message in lace.c
Forgot it was there.
Since an actual LICENSE file was added on July 13, that should supercede
the accidental public domain declaration for changes to lace.c.
But yes, the lace.c code on July 12, 2015 is public domain.
This change should not be an issue anyway since the ISC license is so
permissive, and the other tools have never been marked as public domain.