Add function getline_LaceX() to lace.h
and test it with a variety of chunk sizes and flush sizes.

This didn't give much insight into why getline_XFile() is misbehaving,
but I bet it has to do with accounting for null terminators.
LaceX puts NUL after the end of the array (though still in the allocated
memory), whereas XFile includes it in the array.

https://github.com/grencez/lace/issues/17