Prefer $(XA this) over $(this) syntax for strings
That preference is communicated a warning for now.
I was going to hard-remove it, but the old syntax led to an infinite
loop, so that's a whole investigation. We'll get there eventually.

Examples and test scripts have been updated with the following commands.

* Find: `find -regex '.*[.]lace$' | xargs -d '\n' grep -H -E -e '\$\([^ ]*\)'`
* Replace: `find -regex '.*[.]lace$' | xargs -d '\n' sed -i -E -e 's/\$\(([^ ]*)\)/$(XA \1)/g'`

Fixes https://github.com/grencez/lace/issues/33